gpt_from_scratch/requirements.txt
torch
transformers
datasets
python-dotenv
accelerate
